+++ Mandelbrot/IGenerator.cs
@@ -7,6 +7,6 @@ namespace Baron.Mandelbrot
 {
     interface IGenerator
     {      
-        byte[] Generate(int aWidth, int aHeight, float aSetStartX, float aSetWidth, float aSetStartY, float aSetHeight, byte[] aPalette);
+        byte[] Generate(int aWidth, int aHeight, double aSetStartX, double aSetWidth, double aSetStartY, double aSetHeight, byte[] aPalette);
     }
 }
